chore: split rust into modules
