+++ bindings/src/skinned/mod.rs
@@ -1,0 +1,1 @@
+pub mod export;
